bridge-cassa/requirements.txt
flask==3.0.*
requests==2.32.*
python-dotenv==1.0.*
waitress==3.0.*
